+++ app/src/components/PostVisualBacktest.js
@@ -7,47 +7,12 @@ import moment from 'moment'
 import { HistoryPriceGQL } from './../pages/Detailed/DetailedGQL'
 import PercentChanges from './PercentChanges'
 import PostVisualBacktestChart from './PostVisualBacktestChart'
+import { binarySearchHistoryPriceIndex } from '../utils/utils'
 
 const getChanges = (start, last, prop = 'priceUsd') =>
   (last[`${prop}`] - start[`${prop}`]) / start[`${prop}`] * 100
 
 const isTotalMarket = ticker => ticker === 'Crypto Market'
-
-const binarySearchDirection = {
-  MOVE_STOP_TO_LEFT: -1,
-  MOVE_START_TO_RIGHT: 1
-}
-
-const isCurrentDatetimeBeforeTarget = (current, target) =>
-  moment(current.datetime).isBefore(moment(target))
-
-const binarySearchHistoryPriceIndex = (history, targetDatetime) => {
-  let start = 0
-  let stop = history.length - 1
-  let middle = Math.floor((start + stop) / 2)
-  while (start < stop) {
-    const searchResult = isCurrentDatetimeBeforeTarget(
-      history[middle],
-      targetDatetime
-    )
-      ? binarySearchDirection.MOVE_START_TO_RIGHT
-      : binarySearchDirection.MOVE_STOP_TO_LEFT
-
-    if (searchResult === binarySearchDirection.MOVE_START_TO_RIGHT) {
-      start = middle + 1
-    } else {
-      stop = middle - 1
-    }
-
-    middle = Math.floor((start + stop) / 2)
-  }
-  // Correcting the result to the first data of post's creation date
-  while (!isCurrentDatetimeBeforeTarget(history[middle], targetDatetime)) {
-    middle--
-  }
-
-  return middle
-}
 
 const propTypes = {
   ticker: PropTypes.string.isRequired,
@@ -100,16 +65,8 @@ const enhance = compose(
     const { historyPrice } = history
     if (!historyPrice || historyPrice.length === 0) return {}
 
-    console.time('history-search')
-    // const start =
-    //   historyPrice[
-    //     historyPrice.findIndex(item =>
-    //       moment(item.datetime).utc().isAfter(moment(updatedAt))
-    //     ) - 1
-    //   ]
     const start =
       historyPrice[binarySearchHistoryPriceIndex(historyPrice, updatedAt)]
-    console.timeEnd('history-search')
 
     const last = historyPrice[historyPrice.length - 1]
     if (!start || !last) return {}
